bug fix. company-vacancy service.

+++ mango-company-service/src/main/java/com/mango/mangocompanyservice/service/CompanyService.java
@@ -5,6 +5,8 @@ import com.mango.mangocompanyservice.dto.company.CompanyInfoDto;
 import com.mango.mangocompanyservice.dto.company.CompanySaveDto;
 import com.mango.mangocompanyservice.dto.company.CompanySearchDto;
 import com.mango.mangocompanyservice.entity.Company;
+import com.mango.mangocompanyservice.exception.AccessForbiddenException;
+import com.mango.mangocompanyservice.exception.BadRequestException;
 import com.mango.mangocompanyservice.exception.CompanyNotFoundException;
 import com.mango.mangocompanyservice.repository.CompanyRepository;
 import lombok.RequiredArgsConstructor;
@@ -27,19 +29,21 @@ public class CompanyService {
     private final UserRedisService userRedisService;
     private final BucketService bucketService;
 
-    public Mono<Response> saveCompany(ServerWebExchange exchange, CompanySaveDto dto) {
+    public Mono<Response<String>> saveCompany(ServerWebExchange exchange, CompanySaveDto dto) {
         return userRedisService.buildUser(exchange)
                 .flatMap(user -> {
                     Company company = buildCompany(user.getId(), dto);
+
                     return companyRepository.save(company)
-                            .then(Mono.just(Response.builder()
+                            .then(Mono.just(Response.<String>builder()
+                                    .code(HttpStatus.OK.value())
                                     .message("Company saved successfully.")
-                                    .status(HttpStatus.CREATED)
+                                    .body("Company saved successfully.")
                                     .build()));
                 })
                 .onErrorResume(e -> {
                     log.error("Company didn't saved: {}", e.getMessage());
-                    return Mono.error(new IllegalArgumentException("Company didn't saved."));
+                    return Mono.error(new BadRequestException(e.getMessage()));
                 });
     }
 
@@ -49,73 +53,80 @@ public class CompanyService {
                 .switchIfEmpty(Mono.error(new CompanyNotFoundException("Company not found.")));
     }
 
-    public Mono<Response> deleteCompany(ServerWebExchange exchange, long id) {
+    public Mono<Response<String>> deleteCompany(ServerWebExchange exchange, long id) {
         return userRedisService.buildUser(exchange)
                 .flatMap(user -> companyRepository.findById(id)
                         .flatMap(company -> {
                             if (user.getId() != company.getCeoId()) {
                                 log.error("You cannot delete other users' companies.");
-                                return Mono.error(new IllegalArgumentException("You cannot delete other users' companies."));
+                                return Mono.error(new AccessForbiddenException("You cannot delete other users' companies."));
                             }
 
-                            bucketService.delete(company.getLogo());
+                            if (company.getLogo() != null && !company.getLogo().isEmpty()) {
+                                bucketService.delete(company.getLogo());
+                            }
 
                             return companyRepository.deleteById(id)
-                                    .then(Mono.just(Response.builder()
+                                    .then(Mono.just(Response.<String>builder()
+                                            .code(HttpStatus.OK.value())
                                             .message("Company deleted successfully.")
-                                            .status(HttpStatus.NO_CONTENT)
+                                            .body("Company deleted successfully.")
                                             .build()));
-                        }))
-                .onErrorResume(e -> {
-                    log.error("Company not found: {}", e.getMessage());
-                    return Mono.error(new CompanyNotFoundException("Company didn't saved."));
-                });
+                        })
+                        .switchIfEmpty(Mono.error(new CompanyNotFoundException("Company not found."))))
+                .onErrorResume(e -> Mono.error(new BadRequestException(e.getMessage())));
     }
 
     public Flux<CompanySearchDto> searchByCompanyName(String name, Pageable pageable) {
         return companyRepository.searchByCompanyName(name, pageable)
-                .switchIfEmpty(Mono.error(new CompanyNotFoundException("Company not found.")));
+                .switchIfEmpty(Flux.empty());
     }
 
-    public Mono<Response> updateWebsite(ServerWebExchange exchange, long id, String website) {
+    public Mono<Response<String>> updateWebsite(ServerWebExchange exchange, long id, String website) {
         return userRedisService.buildUser(exchange)
                 .flatMap(user -> companyRepository.findById(id)
                         .flatMap(company -> {
                             if (user.getId() != company.getCeoId()) {
-                                return Mono.error(new IllegalArgumentException("You cannot edit other users' companies."));
+                                return Mono.error(new AccessForbiddenException("You cannot edit other users' companies."));
                             }
 
                             return companyRepository.updateWebsite(id, website)
-                                    .then(Mono.just(Response.builder()
+                                    .then(Mono.just(Response.<String>builder()
+                                            .code(HttpStatus.OK.value())
                                             .message("Company website link updated successfully.")
-                                            .status(HttpStatus.OK)
+                                            .body(website)
                                             .build()));
-                        }))
+                        })
+                        .switchIfEmpty(Mono.error(new CompanyNotFoundException("Company not found."))))
                 .onErrorResume(e -> {
                     log.error("Company not found: {}", e.getMessage());
-                    return Mono.error(new CompanyNotFoundException("Company not found."));
+                    return Mono.error(new BadRequestException(e.getMessage()));
                 });
     }
 
-    public Mono<Response> updateLogo(ServerWebExchange exchange, long id, FilePart file) {
+    public Mono<Response<String>> updateLogo(ServerWebExchange exchange, long id, FilePart file) {
         return userRedisService.buildUser(exchange)
                 .flatMap(user -> companyRepository.findById(id)
                         .flatMap(company -> {
-                            if (user.getId() != company.getCeoId()) {
-                                return Mono.error(new IllegalArgumentException("You cannot edit other users' companies."));
+                            if (!file.filename().matches(".*\\.(png|jpg|jpeg)$")) {
+                                return Mono.error(new BadRequestException("The logo format is not supported. " +
+                                                                            "Please select a png, jpg file."));
                             }
 
-                            String logo = bucketService.save(file);
-                            return companyRepository.updateLogo(id, logo)
-                                    .then(Mono.just(Response.builder()
-                                            .message("Company website logo updated successfully.")
-                                            .status(HttpStatus.OK)
-                                            .build()));
-                        }))
-                .onErrorResume(e -> {
-                    log.error("Company not found: {}", e.getMessage());
-                    return Mono.error(new CompanyNotFoundException("Company not found."));
-                });
+                            if (user.getId() != company.getCeoId()) {
+                                return Mono.error(new AccessForbiddenException("You cannot edit other users' companies."));
+                            }
+
+                            return bucketService.save(file)
+                                    .flatMap(logo -> companyRepository.updateLogo(id, logo)
+                                            .then(Mono.just(Response.<String>builder()
+                                                    .code(HttpStatus.OK.value())
+                                                    .message("Company website logo updated successfully.")
+                                                    .body(logo)
+                                                    .build())));
+                        })
+                        .switchIfEmpty(Mono.error(new CompanyNotFoundException("Company not found."))))
+                .onErrorResume(e -> Mono.error(new BadRequestException(e.getMessage())));
     }
 
     private CompanyInfoDto buildCompanyInfo(Company company) {
